obsolete method toAscii replaced with toLatin1; replaced hidden variables; eliminated unnecessary variables; added channel description to inbox; corrected minor bugs in inbox; corrected some other bugs, like direction of cabytecontroller

+++ caQtDM_QtControls/src/cabyte.cpp
@@ -87,7 +87,7 @@ bool caByte::bitState(long value, int bitNr)
 
 void caByte::setValue(long value)
 {
-    thisValue = (long)value;
+    thisValue = value;
     drawByte(thisValue, thisTrueColor, thisFalseColor);
 }
